+++ common/domain/src/main/java/com/cardinalblue/domain/Author.kt
@@ -4,5 +4,5 @@ data class Author(
     val name: String,
     val username: String,
     val avatarPath: String?,
-    val rating: Int?
+    val rating: Float?
 )
